requirements/base.txt
Django==5.1.8
djangorestframework==3.16.0
django-filter==24.2
django-environ==0.11.2
psycopg2-binary==2.9.9
mergedeep==1.3.4
